Better code reuse between repls for scala 2.10 and 2.11
