+++ repl/src/main/scala_2.10/org/apache/spark/repl/h2o/H2OIMain.scala
@@ -26,6 +26,7 @@ import org.apache.spark.repl.{Main, SparkIMain}
 
 import scala.collection.mutable
 import scala.tools.nsc.Settings
+import scala.tools.nsc.interpreter.IMain.naming
 
 /**
   * SparkIMain allowing multiple interpreters to coexist in parallel. Each line in repl is wrapped in a package which
@@ -34,64 +35,21 @@ import scala.tools.nsc.Settings
 private[repl] class H2OIMain private(initialSettings: Settings,
                                      interpreterWriter: IntpResponseWriter,
                                      val sessionID: Int,
-                                     propagateExceptions: Boolean = false) extends{
+                                     propagateExceptions: Boolean = false) extends {
   override private[repl] val outputDir: File = H2OIMain.classOutputDirectory
-} with SparkIMain(initialSettings, interpreterWriter, propagateExceptions) {
+} with SparkIMain(initialSettings, interpreterWriter, propagateExceptions) with H2OIMainHelper {
 
-  setupClassNames()
+  setupClassNames(naming, sessionId)
 
-  /**
-    * Ensure that each class defined in repl is in a package containing number of repl session
-    */
-  def setupClassNames() = {
-    import naming._
-    // sessionNames is lazy val and needs to be accessed first in order to be then set again to our desired value
-    naming.sessionNames.line
-    val fieldSessionNames = naming.getClass.getDeclaredField("sessionNames")
-    fieldSessionNames.setAccessible(true)
-    fieldSessionNames.set(naming, new SessionNames {
-      override def line  = "intp_id_" + sessionID + propOr("line")
-    })
-  }
 }
 
 
-object H2OIMain {
+object H2OIMain extends H2OIMainHelper {
 
   val existingInterpreters = mutable.HashMap.empty[Int, H2OIMain]
-  private var interpreterClassloader: InterpreterClassLoader = _
-  private var _initialized = false
-
-  private def setClassLoaderToSerializers(classLoader: ClassLoader): Unit = {
-    SparkEnv.get.serializer.setDefaultClassLoader(classLoader)
-    SparkEnv.get.closureSerializer.setDefaultClassLoader(classLoader)
-  }
-
-  private def initialize(sc: SparkContext): Unit = {
-    if (Main.interp != null) {
-      // Application has been started using SparkShell script.
-      // Set the original interpreter classloader as the fallback class loader for all
-      // class not defined in our custom REPL.
-      interpreterClassloader = new InterpreterClassLoader(Main.interp.intp.classLoader)
-    } else {
-      // Application hasn't been started using SparkShell.
-      // Set the context classloader as the fallback class loader for all
-      // class not defined in our custom REPL
-      interpreterClassloader = new InterpreterClassLoader(Thread.currentThread.getContextClassLoader)
-    }
-    setClassLoaderToSerializers(interpreterClassloader)
-  }
-
-  def getInterpreterClassloader: InterpreterClassLoader = {
-    interpreterClassloader
-  }
-
 
   def createInterpreter(sc: SparkContext, settings: Settings, interpreterWriter: IntpResponseWriter, sessionId: Int): H2OIMain = synchronized {
-    if(!_initialized){
-      initialize(sc)
-      _initialized = true
-    }
+    initializeClassLoader(sc)
     existingInterpreters += (sessionId -> new H2OIMain(settings, interpreterWriter, sessionId, false))
     existingInterpreters(sessionId)
   }
@@ -99,15 +57,10 @@ object H2OIMain {
   private lazy val _classOutputDirectory = {
     if (Main.interp != null) {
       // Application was started using shell, we can reuse this directory
-     Main.interp.intp.getClassOutputDirectory
+      Main.interp.intp.getClassOutputDirectory
     } else {
-      // REPL hasn't been started yet, create a new directory
-      val conf = new SparkConf()
-      val rootDir = conf.getOption("spark.repl.classdir").getOrElse(Utils.getLocalDir(conf))
-      val outputDir = Utils.createTempDir(root = rootDir, namePrefix = "repl")
-      outputDir
+      newREPLDirectory()
     }
   }
-
   def classOutputDirectory = _classOutputDirectory
 }
